pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Jun",email = "dlwnsehd13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython (>=9.5.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
